fix(ui,app): clear the lint debt inherited from the studio work
The quality gates lint the whole workspace rather than the diff, so this
branch cannot pass them until upstream's own violations are gone: two
`option_if_let_else` errors in the kithara-ui atoms, 17 ast-grep errors
across three rules, and 10 new arch violations.

Paid down in code, with no suppression and no baseline growth:
private module-level constants moved into the impl or function that owns
them; `StudioRegistry::new` replaced by `Default`; `Binding` collapsed
from four identically-shaped enum variants into a struct with a
`BindingKind` tag, dropping its hand-written `id()` / `key()` accessors;
`ControlNode::size` returns a reference like its sibling `bindings()`;
`gui/studio_ui/reads/` moved to `gui/studio_reads/` for module depth;
`widgets/mini_wave.rs` to `widgets/wave/mini.rs` and
`widgets/{chrome,layout_preview}.rs` to `widgets/module/` for directory
width; `doc/skin.rs` split into `doc/skin/{document,primitives,controls,
panels}.rs`; the unused `DeckSet::set_group_master` deleted.
