+++ crates/kithara-app/src/gui/mod.rs
@@ -3,6 +3,7 @@ mod deck;
 mod frontend;
 mod message;
 mod mix;
+mod studio_reads;
 mod studio_ui;
 mod subscription;
 mod theme;
