+++ crates/kithara-app/src/gui/studio_reads/deck.rs
