+++ crates/kithara-app/src/gui/studio_reads/value.rs
