+++ crates/kithara-app/src/gui/studio_ui/scope.rs
@@ -1,7 +1,7 @@
 /// Channel letter -> session deck position. The studio addresses decks by
 /// their channel letter, in control paths and in binding scopes alike, and the
 /// letter is the deck's position in the session.
-pub(super) fn deck_index(letter: &str) -> Option<usize> {
+pub(in crate::gui) fn deck_index(letter: &str) -> Option<usize> {
     let [byte] = letter.as_bytes() else {
         return None;
     };
