+++ crates/kithara-play/src/rt/limiter.rs
@@ -13,14 +13,10 @@ use firewheel::{
 use kithara_audio::PeakLimiter;
 use kithara_test_utils::kithara;
 
-const SESSION_CEILING: f32 = 0.98;
-
-const SESSION_RELEASE_MS: f32 = 50.0;
-
-const STEREO: NonZeroUsize = NonZeroUsize::new(2).expect("2 is non-zero");
-
-const _: () = assert!(SESSION_CEILING > 0.0 && SESSION_CEILING <= 1.0);
-const _: () = assert!(SESSION_RELEASE_MS > 0.0);
+const _: () = {
+    assert!(LimiterProcessor::CEILING > 0.0 && LimiterProcessor::CEILING <= 1.0);
+    assert!(LimiterProcessor::RELEASE_MS > 0.0);
+};
 
 /// Firewheel adapter around the shared [`PeakLimiter`]: the only site that sees
 /// Firewheel buffers.
@@ -53,23 +49,27 @@ struct LimiterProcessor {
 }
 
 impl LimiterProcessor {
+    const CEILING: f32 = 0.98;
+    const RELEASE_MS: f32 = 50.0;
+    const STEREO: NonZeroUsize = NonZeroUsize::new(2).expect("2 is non-zero");
+
     fn new(sample_rate: NonZeroU32) -> Self {
         Self {
-            limiter: build_limiter(sample_rate),
+            limiter: Self::build_limiter(sample_rate),
         }
     }
-}
 
-fn build_limiter(sample_rate: NonZeroU32) -> PeakLimiter {
-    // The session constants are const-asserted valid above, so construction
-    // cannot fail.
-    PeakLimiter::new(sample_rate, STEREO, SESSION_CEILING, SESSION_RELEASE_MS)
-        .expect("session limiter constants are const-asserted valid")
+    fn build_limiter(sample_rate: NonZeroU32) -> PeakLimiter {
+        // The session constants are const-asserted valid above, so construction
+        // cannot fail.
+        PeakLimiter::new(sample_rate, Self::STEREO, Self::CEILING, Self::RELEASE_MS)
+            .expect("session limiter constants are const-asserted valid")
+    }
 }
 
 impl AudioNodeProcessor for LimiterProcessor {
     fn new_stream(&mut self, stream_info: &StreamInfo, _context: &mut ProcStreamCtx) {
-        self.limiter = build_limiter(stream_info.sample_rate);
+        self.limiter = Self::build_limiter(stream_info.sample_rate);
     }
 
     #[kithara::rtsan_forbid_blocking]
@@ -80,14 +80,14 @@ impl AudioNodeProcessor for LimiterProcessor {
         _events: &mut ProcEvents,
         _extra: &mut ProcExtra,
     ) -> ProcessStatus {
-        if buffers.inputs.len() < STEREO.get() || buffers.outputs.len() < STEREO.get() {
+        if buffers.inputs.len() < Self::STEREO.get() || buffers.outputs.len() < Self::STEREO.get() {
             return ProcessStatus::Bypass;
         }
 
         buffers.outputs[0][..info.frames].copy_from_slice(&buffers.inputs[0][..info.frames]);
         buffers.outputs[1][..info.frames].copy_from_slice(&buffers.inputs[1][..info.frames]);
 
-        if info.in_silence_mask.all_channels_silent(STEREO.get()) {
+        if info.in_silence_mask.all_channels_silent(Self::STEREO.get()) {
             return ProcessStatus::OutputsModifiedWithMask(MaskType::Silence(info.in_silence_mask));
         }
