+++ crates/kithara-ui/src/doc/module.rs
@@ -613,7 +613,7 @@ pub enum ControlNode {
 }
 
 impl ControlNode {
-    pub(crate) fn size(&self) -> Option<SizeSpec> {
+    pub(crate) fn size(&self) -> Option<&SizeSpec> {
         match self {
             Self::Include { .. } => None,
             Self::Row { size, .. }
@@ -655,7 +655,7 @@ impl ControlNode {
             | Self::Chip { size, .. }
             | Self::Knob { size, .. }
             | Self::VuStereo { size, .. }
-            | Self::VuVertical { size, .. } => *size,
+            | Self::VuVertical { size, .. } => size.as_ref(),
         }
     }
