--- crates/kithara-ui/src/doc/skin.rs
@@ -1,1046 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-use super::ron_io;
-use crate::{
-    envelope::{self, DocKind},
-    error::UiDocError,
-    ids::{DocId, SourceUri},
-    layout::FrameSides,
-    module::WindowControlsStyle,
-    size::SizeSpec,
-};
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct SkinDoc {
-    pub schema: String,
-    pub version: u32,
-    pub id: DocId,
-    pub palette: PaletteDoc,
-    pub layout: LayoutSkin,
-    pub chrome: ChromeSkin,
-    pub window: WindowSkin,
-    pub text_input: TextInputSkin,
-    pub knob: KnobSkin,
-    pub crossfader: CrossfaderSkin,
-    pub vu_stereo: VuStereoSkin,
-    pub vu_vertical: VuVerticalSkin,
-    pub vis: VisSkin,
-    pub toggle: ToggleSkin,
-    pub checkbox: CheckboxSkin,
-    pub readout: ReadoutSkin,
-    pub chip: ChipSkin,
-    pub button: ButtonSkin,
-    pub nav: NavSkin,
-    pub tab_large: TabLargeSkin,
-    pub text: TextSkin,
-    pub segmented: SegmentedSkin,
-    pub select: SelectSkin,
-    pub status_dot: StatusDotSkin,
-    pub swatch: SwatchSkin,
-    pub cell: CellSkin,
-    pub fader: FaderSkin,
-    pub wave: WaveSkin,
-    pub deck: DeckSkin,
-    pub global_bar: GlobalBarSkin,
-    pub divider: DividerSkin,
-    pub drag: DragSkin,
-    pub meter: MeterSkin,
-    pub telemetry: TelemetrySkin,
-    pub tree: TreeSkin,
-    pub track_list: TrackListSkin,
-    pub layout_preview: LayoutPreviewSkin,
-}
-
-#[derive(Clone, Debug, Deserialize, Eq, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct PaletteDoc {
-    pub bg: String,
-    pub bg_deep: String,
-    pub bg_inset: String,
-    pub bg_panel: String,
-    pub bg_footer: String,
-    pub bg_panel_2: String,
-    pub bg_select: String,
-    pub line: String,
-    pub line_dim: String,
-    pub line_inner: String,
-    pub line_soft: String,
-    pub text: String,
-    pub text_dim: String,
-    pub muted: String,
-    pub accent: String,
-    pub accent_strong: String,
-    pub accent_soft: String,
-    pub danger: String,
-    pub success: String,
-    pub warning: String,
-    pub wave_low: String,
-    pub wave_mid: String,
-    pub wave_high: String,
-}
-
-impl PaletteDoc {
-    fn validate(&self, origin: &SourceUri) -> Result<(), UiDocError> {
-        for value in [
-            &self.bg,
-            &self.bg_deep,
-            &self.bg_inset,
-            &self.bg_panel,
-            &self.bg_footer,
-            &self.bg_panel_2,
-            &self.bg_select,
-            &self.line,
-            &self.line_dim,
-            &self.line_inner,
-            &self.line_soft,
-            &self.text,
-            &self.text_dim,
-            &self.muted,
-            &self.accent,
-            &self.accent_strong,
-            &self.accent_soft,
-            &self.danger,
-            &self.success,
-            &self.warning,
-            &self.wave_low,
-            &self.wave_mid,
-            &self.wave_high,
-        ] {
-            parse_color(value, origin)?;
-        }
-        Ok(())
-    }
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, Eq, PartialEq, Serialize)]
-#[non_exhaustive]
-pub enum ColorRole {
-    Bg,
-    BgDeep,
-    BgInset,
-    BgPanel,
-    BgFooter,
-    BgPanel2,
-    BgSelect,
-    Line,
-    LineDim,
-    LineInner,
-    LineSoft,
-    Text,
-    TextDim,
-    Muted,
-    Accent,
-    AccentStrong,
-    AccentSoft,
-    Danger,
-    Success,
-    Warning,
-    WaveLow,
-    WaveMid,
-    WaveHigh,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, Eq, PartialEq, Serialize)]
-#[non_exhaustive]
-pub enum FontWeight {
-    Normal,
-    Medium,
-    Semibold,
-    Bold,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, Eq, PartialEq, Serialize)]
-#[non_exhaustive]
-pub enum FontFamily {
-    Display,
-    Sans,
-    Mono,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct FontSkin {
-    pub size: f32,
-    pub weight: FontWeight,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct TextRoleSkin {
-    pub font: FontFamily,
-    pub weight: FontWeight,
-    pub size: f32,
-    pub spacing: f32,
-    pub color: ColorRole,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct FrameSkin {
-    pub radius: f32,
-    pub border_width: f32,
-    pub border: ColorRole,
-}
-
-/// Scale beside a fader: hairlines with a longer, brighter one at centre.
-/// `thickness` runs along the scale, `length` across it, whatever the axis.
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct TickSkin {
-    pub count: usize,
-    pub thickness: f32,
-    pub length: f32,
-    pub center_length: f32,
-    pub gap: f32,
-    pub inset: f32,
-    pub color: ColorRole,
-    pub center_color: ColorRole,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct LayoutSkin {
-    pub grid_gap: f32,
-    pub grid_pad: f32,
-    pub fill_weight_scale: f32,
-    pub fill_weight_min: f32,
-    pub size_gap: f32,
-    pub size_pad: f32,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct ChromeSkin {
-    pub frame: FrameSkin,
-    pub secondary_frame: FrameSkin,
-    pub header_frame: FrameSkin,
-    pub chip_frame: FrameSkin,
-    pub title_frame: FrameSkin,
-    pub chevron_frame: FrameSkin,
-    pub footer_frame: FrameSkin,
-    pub header_height: f32,
-    pub chip_pad: f32,
-    pub chip_text_size: f32,
-    pub title_text_size: f32,
-    pub chevron_size: f32,
-    pub chevron_icon_size: f32,
-    pub chevron_stroke_width: f32,
-    pub footer_height: f32,
-    pub footer_pad: f32,
-    pub footer_text_size: f32,
-    pub inner_line_width: f32,
-    pub panel_background: ColorRole,
-    pub header_background: ColorRole,
-    pub title_background: ColorRole,
-    pub chip_background: ColorRole,
-    pub chip_text: ColorRole,
-    pub title_text: ColorRole,
-    pub chevron_color: ColorRole,
-    pub footer_background: ColorRole,
-    pub footer_text: ColorRole,
-    pub inner_line: ColorRole,
-    pub corner_size: f32,
-    pub corner_width: f32,
-    pub corner_offset: f32,
-    pub corner_color: ColorRole,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct WindowSkin {
-    pub titlebar_height: f32,
-    /// Thickness of the drag zones framing a window that draws its own chrome.
-    pub resize_edge: f32,
-    pub titlebar_padding_x: f32,
-    pub titlebar_text: TextRoleSkin,
-    pub icon_color: ColorRole,
-    pub icon_hover_color: ColorRole,
-    pub icon_stroke_width: f32,
-    pub standard_minus_icon_size: f32,
-    pub standard_maximize_icon_size: f32,
-    pub standard_close_icon_size: f32,
-    pub standard_gap: f32,
-    pub standard_padding: f32,
-    pub compact_minus_icon_size: f32,
-    pub compact_maximize_icon_size: f32,
-    pub compact_close_icon_size: f32,
-    pub compact_gap: f32,
-    pub compact_padding: f32,
-    pub close_wide_cell_size: f32,
-    pub close_wide_icon_size: f32,
-    pub close_wide_divider_width: f32,
-    pub close_wide_divider_color: ColorRole,
-    pub close_micro_cell_size: f32,
-    pub close_micro_icon_size: f32,
-    pub close_framed_cell_size: f32,
-    pub close_framed_icon_size: f32,
-    pub close_framed_frame: FrameSkin,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub(crate) enum WindowControlSkin {
-    Buttons {
-        minus_icon_size: f32,
-        maximize_icon_size: f32,
-        close_icon_size: f32,
-        gap: f32,
-        padding: f32,
-    },
-    Close {
-        cell_size: f32,
-        icon_size: f32,
-        frame: Option<FrameSkin>,
-        divider: Option<(f32, ColorRole)>,
-    },
-}
-
-impl WindowSkin {
-    pub(crate) const fn controls(self, style: WindowControlsStyle) -> WindowControlSkin {
-        match style {
-            WindowControlsStyle::Standard => WindowControlSkin::Buttons {
-                minus_icon_size: self.standard_minus_icon_size,
-                maximize_icon_size: self.standard_maximize_icon_size,
-                close_icon_size: self.standard_close_icon_size,
-                gap: self.standard_gap,
-                padding: self.standard_padding,
-            },
-            WindowControlsStyle::Compact => WindowControlSkin::Buttons {
-                minus_icon_size: self.compact_minus_icon_size,
-                maximize_icon_size: self.compact_maximize_icon_size,
-                close_icon_size: self.compact_close_icon_size,
-                gap: self.compact_gap,
-                padding: self.compact_padding,
-            },
-            WindowControlsStyle::CloseWide => WindowControlSkin::Close {
-                cell_size: self.close_wide_cell_size,
-                icon_size: self.close_wide_icon_size,
-                frame: None,
-                divider: Some((self.close_wide_divider_width, self.close_wide_divider_color)),
-            },
-            WindowControlsStyle::CloseMicro => WindowControlSkin::Close {
-                cell_size: self.close_micro_cell_size,
-                icon_size: self.close_micro_icon_size,
-                frame: None,
-                divider: None,
-            },
-            WindowControlsStyle::CloseFramed => WindowControlSkin::Close {
-                cell_size: self.close_framed_cell_size,
-                icon_size: self.close_framed_icon_size,
-                frame: Some(self.close_framed_frame),
-                divider: None,
-            },
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct TextInputSkin {
-    pub height: f32,
-    pub padding_x: f32,
-    pub padding_y: f32,
-    pub font: FontSkin,
-    pub radius: f32,
-    pub border_width: f32,
-    pub idle_border_width: f32,
-    pub border: ColorRole,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct KnobSkin {
-    pub size: SizeSpec,
-    pub body_ratio: f32,
-    pub body_fill: ColorRole,
-    pub body_border: ColorRole,
-    pub track_color: ColorRole,
-    pub value_color: ColorRole,
-    pub indicator_color: ColorRole,
-    pub drag_range: f32,
-    pub wheel_step: f32,
-    pub indicator_width: f32,
-    pub outer_inset: f32,
-    pub start_angle: f32,
-    pub neutral_angle: f32,
-    pub sweep_angle: f32,
-    pub track_width: f32,
-    pub track_alpha: f32,
-    pub body_border_width: f32,
-    pub label_text: TextRoleSkin,
-    pub label_gap: f32,
-    pub label_height: f32,
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct CrossfaderSkin {
-    pub size: SizeSpec,
-    pub padding_x: f32,
-    pub padding_top: f32,
-    pub padding_bottom: f32,
-    pub rail_height: f32,
-    pub rail_frame: FrameSkin,
-    pub rail_background: ColorRole,
-    pub thumb_width: f32,
-    pub thumb_height: f32,
-    pub thumb_color: ColorRole,
-    pub thumb_notch_width: f32,
-    pub thumb_notch_height: f32,
-    pub thumb_notch_color: ColorRole,
-    pub ticks: TickSkin,
-    pub label_text: FontSkin,
-    pub label_color: ColorRole,
-    pub label_gap: f32,
-    pub letter_text: FontSkin,
-    pub letter_color: ColorRole,
-    pub arrow_size: f32,
-    pub arrow_color: ColorRole,
-    pub arrow_gap: f32,
-    pub left_label: String,
-    pub center_label: String,
-    pub right_label: String,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct VuStereoSkin {
-    pub size: SizeSpec,
-    pub carriage_width: f32,
-    pub channel_l_y: f32,
-    pub channel_r_y: f32,
-    pub segment_count: usize,
-    pub segment_gap: f32,
-    pub segment_height: f32,
-    pub segment_width: f32,
-    pub warning_threshold: f32,
-    pub danger_threshold: f32,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct VuVerticalSkin {
-    pub size: SizeSpec,
-    pub fader_width: f32,
-    pub segment_gap: f32,
-    pub segment_height: f32,
-    pub segment_inset_x: f32,
-    pub ticks: TickSkin,
-    pub thumb_height: f32,
-    pub thumb_color: ColorRole,
-    pub thumb_notch_offset: f32,
-    pub thumb_notch_height: f32,
-    pub thumb_notch_color: ColorRole,
-    pub warning_threshold: f32,
-    pub danger_threshold: f32,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct VisSkin {
-    pub size: SizeSpec,
-    pub header_height: f32,
-    pub footer_height: f32,
-    pub nav_cell_size: f32,
-    pub nav_padding_x: f32,
-    pub nav_padding_y: f32,
-    pub name_padding_x: f32,
-    pub index_padding_x: f32,
-    pub footer_padding_x: f32,
-    pub title: TextRoleSkin,
-    pub meta: TextRoleSkin,
-    pub nav_text: FontSkin,
-    pub nav_frame: FrameSkin,
-    pub nav_background: ColorRole,
-    pub nav_text_color: ColorRole,
-    pub icon_size: f32,
-    pub icon_color: ColorRole,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct ToggleSkin {
-    pub size: SizeSpec,
-    pub active_frame: FrameSkin,
-    pub inactive_frame: FrameSkin,
-    pub thumb_size: f32,
-    pub thumb_inset: f32,
-    pub thumb_radius: f32,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct CheckboxSkin {
-    pub size: SizeSpec,
-    pub active_frame: FrameSkin,
-    pub inactive_frame: FrameSkin,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct ReadoutSkin {
-    pub size: SizeSpec,
-    pub frame: FrameSkin,
-    pub padding_x: f32,
-    pub padding_y: f32,
-    pub spacing: f32,
-    pub label: FontSkin,
-    pub value: FontSkin,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct ChipSkin {
-    pub size: SizeSpec,
-    pub active_frame: FrameSkin,
-    pub inactive_frame: FrameSkin,
-    pub padding_x: f32,
-    pub padding_y: f32,
-    pub deck_text: FontSkin,
-    pub routing_text: FontSkin,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct ButtonSkin {
-    pub size: SizeSpec,
-    pub frame: FrameSkin,
-    pub primary_frame: FrameSkin,
-    /// A transport cell draws no border of its own; these sides say where the
-    /// seam between neighbouring cells goes.
-    pub transport_sides: FrameSides,
-    pub padding_x: f32,
-    pub padding_y: f32,
-    pub text: FontSkin,
-    pub primary_text: FontSkin,
-    pub icon_size: f32,
-    pub transport_icon_size: f32,
-    pub icon_gap: f32,
-    pub micro_size: f32,
-    pub micro_icon_size: f32,
-    pub transport_fill: u16,
-    pub primary_fill: u16,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct NavSkin {
-    pub item_height: f32,
-    pub marker_width: f32,
-    pub icon_size: f32,
-    pub text_size: f32,
-    pub text_pad_x: f32,
-    pub icon_gap: f32,
-    pub header_height: f32,
-    pub header_text_size: f32,
-    pub header_icon_size: f32,
-    pub pad_y: f32,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct TabLargeSkin {
-    pub height: f32,
-    pub pad_x: f32,
-    pub pad_y: f32,
-    pub underline_width: f32,
-    pub text_size: f32,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct TextSkin {
-    pub size: SizeSpec,
-    pub brand: TextRoleSkin,
-    pub deck_letter: TextRoleSkin,
-    /// Tone the deck letter takes while its `Text.active` binding reads true.
-    pub deck_letter_active: ColorRole,
-    pub track_title: TextRoleSkin,
-    pub body: TextRoleSkin,
-    pub telemetry: TextRoleSkin,
-    pub micro_label: TextRoleSkin,
-    pub section: TextRoleSkin,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct SegmentedSkin {
-    pub size: SizeSpec,
-    pub frame: FrameSkin,
-    pub background: ColorRole,
-    pub active_background: ColorRole,
-    pub active_text: ColorRole,
-    pub inactive_text: ColorRole,
-    pub padding_x: f32,
-    pub text: FontSkin,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct SelectSkin {
-    pub size: SizeSpec,
-    pub frame: FrameSkin,
-    pub background: ColorRole,
-    pub padding_x: f32,
-    pub padding_y: f32,
-    pub text: FontSkin,
-    pub text_color: ColorRole,
-    pub chevron_size: f32,
-    pub chevron_color: ColorRole,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct StatusDotSkin {
-    pub size: SizeSpec,
-    pub dot_size: f32,
-    pub gap: f32,
-    pub text: FontSkin,
-    pub text_color: ColorRole,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct SwatchSkin {
-    pub size: SizeSpec,
-    pub box_height: f32,
-    pub frame: FrameSkin,
-    pub box_label_gap: f32,
-    pub label_hex_gap: f32,
-    pub label: TextRoleSkin,
-    pub hex: TextRoleSkin,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct CellSkin {
-    pub size: SizeSpec,
-    pub frame: FrameSkin,
-    pub highlighted_frame: FrameSkin,
-    pub background: ColorRole,
-    pub label_gap: f32,
-    pub label_height: f32,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct FaderSkin {
-    pub size: SizeSpec,
-    pub content_gap: f32,
-    pub control_height: f32,
-    pub control_padding_x: f32,
-    pub control_padding_y: f32,
-    pub rail_frame: FrameSkin,
-    pub handle_frame: FrameSkin,
-    pub strip_frame: FrameSkin,
-    pub handle_width: u16,
-    pub handle_color: ColorRole,
-    pub icon_size: f32,
-    pub icon_width: f32,
-    pub label: FontSkin,
-    pub label_width: f32,
-    pub rail_width: f32,
-    pub segment_count: usize,
-    pub segment_gap: f32,
-    pub segment_height: f32,
-    pub slider_height: f32,
-    pub step: f64,
-    pub strip_height: f32,
-    pub strip_padding: f32,
-    pub ticks_height: f32,
-    pub tick_height: f32,
-    pub tick_step: f32,
-    pub tick_width: f32,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct WaveSkin {
-    pub size: SizeSpec,
-    pub frame: FrameSkin,
-    pub overlay: WaveOverlaySkin,
-    pub background: ColorRole,
-    pub bar_gap: f32,
-    /// Width of every band bar; bands nest by level, not by width.
-    pub bar_width: f32,
-    pub content_inset: f32,
-    pub cue_badge_background: ColorRole,
-    pub cue_badge_size: f32,
-    pub cue_badge_text: FontSkin,
-    pub cue_badge_text_color: ColorRole,
-    pub cue_line_width: f32,
-    pub downbeat_alpha: f32,
-    pub grid_alpha: f32,
-    pub grid_width: f32,
-    pub loop_bound_width: f32,
-    pub loop_fill_alpha: f32,
-    pub played_alpha: f32,
-    /// Overview strips dim harder than the hero wave.
-    pub overview_played_alpha: f32,
-    pub playhead_marker_height: f32,
-    pub playhead_marker_width: f32,
-    pub playhead_width: f32,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct WaveOverlaySkin {
-    pub background: ColorRole,
-    pub background_alpha: f32,
-    pub height: f32,
-    pub padding_x: f32,
-    pub padding_y: f32,
-    pub gap: f32,
-    pub art_frame: FrameSkin,
-    pub art_background: ColorRole,
-    pub art_size: f32,
-    pub art_label: FontSkin,
-    pub art_label_color: ColorRole,
-    pub title: FontSkin,
-    pub title_color: ColorRole,
-    pub artist: FontSkin,
-    pub artist_color: ColorRole,
-    pub summary_gap: f32,
-    pub readout_frame: FrameSkin,
-    pub readout_background: ColorRole,
-    pub bpm_width: f32,
-    pub key_width: f32,
-    pub readout_height: f32,
-    pub readout_padding_x: f32,
-    pub readout_padding_y: f32,
-    pub readout_gap: f32,
-    pub readout_label: FontSkin,
-    pub readout_label_color: ColorRole,
-    pub readout_value: FontSkin,
-    pub bpm_color: ColorRole,
-    pub key_color: ColorRole,
-    pub remain_width: f32,
-    pub remain_color: ColorRole,
-    pub badge_frame: FrameSkin,
-    pub badge_background: ColorRole,
-    pub badge_size: f32,
-    pub badge_text: FontSkin,
-    pub badge_text_color: ColorRole,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct DeckSkin {
-    pub summary_size: SizeSpec,
-    pub bpm_size: SizeSpec,
-    pub time_size: SizeSpec,
-    pub artist: FontSkin,
-    pub bpm_text: FontSkin,
-    pub micro_source: FontSkin,
-    pub micro_summary_gap: f32,
-    pub micro_title: FontSkin,
-    pub readout_gap: f32,
-    pub readout_label: FontSkin,
-    pub summary_fill: u16,
-    pub summary_height: f32,
-    pub summary_padding_x: f32,
-    pub summary_padding_y: f32,
-    pub time_padding_x: f32,
-    pub time_padding_y: f32,
-    pub time_text: FontSkin,
-    pub title: FontSkin,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct GlobalBarSkin {
-    pub brand_size: SizeSpec,
-    pub spacer_size: SizeSpec,
-    pub preset_size: SizeSpec,
-    pub settings_size: SizeSpec,
-    pub selector_frame: FrameSkin,
-    pub chip_frame: FrameSkin,
-    pub settings_frame: FrameSkin,
-    pub brand_gap: f32,
-    pub brand_padding_x: f32,
-    pub brand_padding_y: f32,
-    pub brand_text: FontSkin,
-    pub chip_gap: f32,
-    pub chip_padding_x: f32,
-    pub chip_padding_y: f32,
-    pub chip_text: FontSkin,
-    pub gear_size: f32,
-    pub height: f32,
-    pub settings_padding: f32,
-    pub selector_padding_x: f32,
-    pub selector_padding_y: f32,
-    pub brand_width: f32,
-    pub selector_width: f32,
-}
-
-/// Horizontal fill bar reporting one scalar, as the design's CPU cell draws it:
-/// an inset track with a hairline frame, filled from the left.
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct MeterSkin {
-    pub size: SizeSpec,
-    pub frame: FrameSkin,
-    pub background: ColorRole,
-    pub fill: ColorRole,
-}
-
-/// Hairline between adjacent cells or control sections.
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct DividerSkin {
-    pub width: f32,
-    pub color: ColorRole,
-}
-
-/// The label the pointer carries while it drags an item.
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct DragSkin {
-    pub frame: FrameSkin,
-    pub background: ColorRole,
-    pub text: TextRoleSkin,
-    pub width: f32,
-    pub height: f32,
-    pub pad_x: f32,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct TelemetrySkin {
-    pub size: SizeSpec,
-    pub frame: FrameSkin,
-    pub padding_x: f32,
-    pub padding_y: f32,
-    pub percent_precision: usize,
-    pub percent_scale: f64,
-    pub percent_width: usize,
-    pub scalar_precision: usize,
-    pub text: FontSkin,
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct TreeSkin {
-    pub size: SizeSpec,
-    pub search_placeholder: String,
-    pub search_height: f32,
-    pub search_icon_width: f32,
-    pub search_icon_size: f32,
-    pub search_padding_x: f32,
-    pub search_text: FontSkin,
-    pub search_background: ColorRole,
-    pub search_divider: ColorRole,
-    pub panel_background: ColorRole,
-    pub panel_padding_top: f32,
-    pub panel_padding_bottom: f32,
-    pub scrollbar_width: f32,
-    pub scrollbar_margin: f32,
-    pub scroller_color: ColorRole,
-    pub scrollbar_background: ColorRole,
-    pub row_height: f32,
-    pub row_padding_right: f32,
-    pub indent_base: f32,
-    pub indent_step: f32,
-    pub marker_width: f32,
-    pub chevron_width: f32,
-    pub chevron_size: f32,
-    pub icon_size: f32,
-    pub content_gap: f32,
-    pub label_text: FontSkin,
-    pub count_text: FontSkin,
-    pub context_height: f32,
-    pub context_padding_x: f32,
-    pub context_icon_size: f32,
-    pub context_gap: f32,
-    pub context_text: FontSkin,
-    pub context_background: ColorRole,
-    pub context_divider: ColorRole,
-    pub context_divider_width: f32,
-    pub scope_text: FontSkin,
-    pub scope_gap: f32,
-    pub scope_padding_x: f32,
-    pub scope_frame: FrameSkin,
-    pub scope_background: ColorRole,
-    pub scope_text_color: ColorRole,
-    pub scope_chevron_size: f32,
-    pub scope_chevron_color: ColorRole,
-    pub scope_menu_background: ColorRole,
-    pub scope_menu_frame: FrameSkin,
-    pub scope_item_height: f32,
-    pub scope_menu_text: ColorRole,
-    pub scope_selected_background: ColorRole,
-    pub scope_selected_text: ColorRole,
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct TrackListSkin {
-    pub size: SizeSpec,
-    pub labels: TrackListLabels,
-    pub row_frame: FrameSkin,
-    pub deck_chip_frame: FrameSkin,
-    pub bpm_badge_frame: FrameSkin,
-    pub bpm_badge_background: ColorRole,
-    pub divider_color: ColorRole,
-    pub divider_hit_width: f32,
-    pub divider_width: f32,
-    pub energy_bar_background: ColorRole,
-    pub cell_padding_x: f32,
-    pub footer_padding_x: f32,
-    pub grid_gap: f32,
-    pub header_height: f32,
-    pub row_height: f32,
-    pub footer_height: f32,
-    pub min_column_width: f32,
-    pub index_width: f32,
-    pub deck_width: f32,
-    pub artist_width: f32,
-    pub bpm_width: f32,
-    pub key_width: f32,
-    pub time_width: f32,
-    pub energy_width: f32,
-    pub transition_width: f32,
-    pub title_min_width: f32,
-    pub scrollbar_background: ColorRole,
-    pub scrollbar_margin: f32,
-    pub scrollbar_width: f32,
-    pub scroller_color: ColorRole,
-    pub deck_chip_width: f32,
-    pub deck_chip_height: f32,
-    pub bpm_badge_height: f32,
-    pub bpm_badge_padding_x: f32,
-    pub energy_bar_width: f32,
-    pub energy_bar_height: f32,
-    pub energy_bar_gap: f32,
-    pub header_text: FontSkin,
-    pub index_text: FontSkin,
-    pub deck_text: FontSkin,
-    pub title_text: FontSkin,
-    pub artist_text: FontSkin,
-    pub bpm_text: FontSkin,
-    pub key_text: FontSkin,
-    pub time_text: FontSkin,
-    pub energy_text: FontSkin,
-    pub transition_text: FontSkin,
-    pub footer_text: FontSkin,
-}
-
-#[derive(Clone, Debug, Deserialize, Eq, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct TrackListLabels {
-    pub index: String,
-    pub deck: String,
-    pub title: String,
-    pub artist: String,
-    pub bpm: String,
-    pub key: String,
-    pub time: String,
-    pub energy: String,
-    pub transition: String,
-    pub footer_tracks: String,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-#[non_exhaustive]
-pub struct LayoutPreviewSkin {
-    pub height: f32,
-    pub line_width: f32,
-    pub module_inset: f32,
-}
-
-/// Parses and validates a complete skin document.
-///
-/// # Errors
-/// Returns [`UiDocError`] when the envelope, body, or palette is invalid.
-pub fn parse_skin(text: &str, origin: &SourceUri) -> Result<SkinDoc, UiDocError> {
-    let envelope = envelope::probe(text, origin)?;
-    if envelope.kind != DocKind::Skin {
-        return Err(UiDocError::WrongDocKind {
-            origin: origin.clone(),
-            expected: DocKind::Skin.name(),
-            found: envelope.kind.name(),
-        });
-    }
-    let document: SkinDoc =
-        ron_io::options()
-            .from_str(text)
-            .map_err(|source| UiDocError::Syntax {
-                origin: origin.clone(),
-                source: Box::new(source),
-            })?;
-    document.palette.validate(origin)?;
-    Ok(document)
-}
-
-pub(crate) fn parse_color(value: &str, origin: &SourceUri) -> Result<[u8; 4], UiDocError> {
-    let digits = value
-        .strip_prefix('#')
-        .ok_or_else(|| bad_color(origin, value))?;
-    if digits.len() != 6 && digits.len() != 8 {
-        return Err(bad_color(origin, value));
-    }
-    let component = |start| {
-        let pair = digits
-            .get(start..start + 2)
-            .ok_or_else(|| bad_color(origin, value))?;
-        u8::from_str_radix(pair, 16).map_err(|_| bad_color(origin, value))
-    };
-    Ok([
-        component(0)?,
-        component(2)?,
-        component(4)?,
-        if digits.len() == 8 {
-            component(6)?
-        } else {
-            255
-        },
-    ])
-}
-
-fn bad_color(origin: &SourceUri, value: &str) -> UiDocError {
-    UiDocError::BadColor {
-        origin: origin.clone(),
-        value: value.to_owned(),
-    }
-}
