+++ crates/kithara-ui/src/doc/skin/mod.rs
@@ -1,0 +1,6 @@
+mod controls;
+mod document;
+mod panels;
+mod primitives;
+
+pub use self::{controls::*, document::*, panels::*, primitives::*};
