+++ crates/kithara-ui/src/expand/mod.rs
@@ -5,5 +5,5 @@ mod node;
 pub(crate) use binding_subst::intern_binding;
 pub use binding_subst::scoped_key;
 pub(crate) use machine::Expander;
-pub use node::{Binding, ControlSpec, DropSpec, ExpandedNode, SurfaceSpec};
+pub use node::{Binding, BindingKind, ControlSpec, DropSpec, ExpandedNode, SurfaceSpec};
 pub(crate) use node::{Budget, ControlSite, ControlVisitor, ExpandedModule};
