+++ crates/kithara-ui/src/widgets/drag_ghost.rs
@@ -15,16 +15,6 @@ use crate::{
     skin::DragSkin,
 };
 
-const ELLIPSIS: char = '\u{2026}';
-
-/// Gap between the pointer and the box it carries, so the box trails the
-/// pointer instead of sitting under it.
-const POINTER_GAP: f32 = 12.0;
-
-/// A canvas cannot measure text, so the label is cut to what fits: the mono
-/// face this widget draws with advances that fraction of its size per glyph.
-const MONO_ADVANCE: f32 = 0.6;
-
 /// What the pointer is carrying, drawn at the pointer over everything the
 /// layout lays out. It paints only: every event passes through it untouched,
 /// and it claims no cursor of its own.
@@ -37,6 +27,16 @@ pub(crate) struct DragGhost {
 }
 
 impl DragGhost {
+    const ELLIPSIS: char = '\u{2026}';
+
+    /// Gap between the pointer and the box it carries, so the box trails the
+    /// pointer instead of sitting under it.
+    const POINTER_GAP: f32 = 12.0;
+
+    /// A canvas cannot measure text, so the label is cut to what fits: the mono
+    /// face this widget draws with advances that fraction of its size per glyph.
+    const MONO_ADVANCE: f32 = 0.6;
+
     pub(crate) fn new(label: Option<String>, skin: &Skin) -> Self {
         let metrics = skin.drag;
         Self {
@@ -51,8 +51,8 @@ impl DragGhost {
     /// The box trails the pointer and stays inside the window.
     fn box_at(&self, pointer: Point, bounds: Rectangle) -> Rectangle {
         let size = Size::new(self.metrics.width, self.metrics.height);
-        let x = (pointer.x + POINTER_GAP).min(bounds.width - size.width);
-        let y = (pointer.y + POINTER_GAP).min(bounds.height - size.height);
+        let x = (pointer.x + Self::POINTER_GAP).min(bounds.width - size.width);
+        let y = (pointer.y + Self::POINTER_GAP).min(bounds.height - size.height);
         Rectangle::new(Point::new(x.max(0.0), y.max(0.0)), size)
     }
 }
@@ -136,7 +136,9 @@ impl canvas::Program<UiEvent> for DragGhost {
 /// Glyphs the box holds between its paddings.
 fn fitting_chars(metrics: DragSkin) -> usize {
     let inner = (metrics.width - metrics.pad_x * 2.0).max(0.0);
-    (inner / (metrics.text.size * MONO_ADVANCE)).trunc().as_()
+    (inner / (metrics.text.size * DragGhost::MONO_ADVANCE))
+        .trunc()
+        .as_()
 }
 
 /// Keeps the label inside the box the skin authored for it.
@@ -147,7 +149,7 @@ fn elide(label: &str, max_chars: usize) -> String {
         return head;
     }
     let mut elided: String = head.chars().take(max_chars.saturating_sub(1)).collect();
-    elided.push(ELLIPSIS);
+    elided.push(DragGhost::ELLIPSIS);
     elided
 }
