+++ crates/kithara-ui/src/widgets/mod.rs
@@ -1,12 +1,10 @@
 pub(crate) mod behavior;
 pub(crate) mod button;
-mod chrome;
 pub(crate) mod deck;
 pub(crate) mod drag_ghost;
 pub(crate) mod fader;
 pub(crate) mod global_bar;
-mod layout_preview;
-pub(crate) mod mini_wave;
+mod module;
 pub(crate) mod nav;
 pub(crate) mod telemetry;
 pub(crate) mod text;
@@ -15,7 +13,7 @@ pub(crate) mod vis;
 pub(crate) mod wave;
 pub(crate) mod wheel;
 pub(crate) mod window;
-pub(crate) use chrome::{DropZone, ModuleChrome, frame_overlay};
-pub use layout_preview::LayoutPreview;
+pub use module::LayoutPreview;
+pub(crate) use module::{DropZone, ModuleChrome, frame_overlay};
 
 pub(crate) use crate::render::event::Widget;
