+++ crates/kithara-ui/src/widgets/module/mod.rs
@@ -1,0 +1,5 @@
+mod chrome;
+mod preview;
+
+pub(crate) use self::chrome::{DropZone, ModuleChrome, frame_overlay};
+pub use self::preview::LayoutPreview;
