+++ crates/kithara-ui/src/widgets/wave/mini.rs
