+++ crates/kithara-ui/src/widgets/wave/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod bars;
 pub(crate) mod hero;
+pub(crate) mod mini;
 pub(crate) mod zoom_math;
