+++ crates/kithara-ui/tests/compile.rs
@@ -5,7 +5,7 @@ use kithara_ui::{
     builtin,
     compile::{CompiledNode, compile},
     error::UiDocError,
-    expand::{Binding, ControlSpec, ExpandedNode},
+    expand::{Binding, BindingKind, ControlSpec, ExpandedNode},
     module::{ChromeStyle, TrackColumn},
     registry::{EndpointCategory, EndpointDesc, ValueKind},
     size::{Dim, SizeSpec},
@@ -80,8 +80,14 @@ fn crossfader_compiles_with_scalar_read_and_write_bindings() {
     };
     let ExpandedNode::Control {
         spec: ControlSpec::Crossfader { ticks: false },
-        read: Some(Binding::Parameter { .. }),
-        write: Some(Binding::Parameter { .. }),
+        read: Some(Binding {
+            kind: BindingKind::Parameter,
+            ..
+        }),
+        write: Some(Binding {
+            kind: BindingKind::Parameter,
+            ..
+        }),
         ..
     } = &**root
     else {
@@ -126,7 +132,10 @@ fn meter_reads_a_scalar_and_refuses_any_other_kind() {
         &**root,
         ExpandedNode::Control {
             spec: ControlSpec::Meter,
-            read: Some(Binding::Telemetry { .. }),
+            read: Some(Binding {
+                kind: BindingKind::Telemetry,
+                ..
+            }),
             ..
         }
     ));
@@ -186,8 +195,14 @@ fn vis_compiles_with_scalar_read_and_select_index_write() {
     };
     let ExpandedNode::Control {
         spec: ControlSpec::Vis,
-        read: Some(Binding::Model { .. }),
-        write: Some(Binding::Parameter { .. }),
+        read: Some(Binding {
+            kind: BindingKind::Model,
+            ..
+        }),
+        write: Some(Binding {
+            kind: BindingKind::Parameter,
+            ..
+        }),
         ..
     } = &**root
     else {
@@ -330,7 +345,12 @@ fn track_list_compiles_typed_columns_and_optional_state_prefix() {
         columns,
         &[TrackColumn::Index, TrackColumn::Title, TrackColumn::Bpm]
     );
-    let Some(Binding::Model { id, .. }) = columns_state else {
+    let Some(Binding {
+        kind: BindingKind::Model,
+        id,
+        ..
+    }) = columns_state
+    else {
         panic!("expected model state prefix");
     };
     assert_eq!(ui.resolve(*id), "ui.tracklist.columns");
@@ -465,7 +485,14 @@ fn module_shell_metadata_compiles_into_the_module_node() {
     assert!(!frame.left);
     assert!(*corners);
     assert_eq!(ui.resolve(*collapsed), "ui.module.deck.collapsed");
-    let Some(Binding::Telemetry { id, key, with }) = footer else {
+    let Some(Binding {
+        kind: BindingKind::Telemetry,
+        id,
+        key,
+        with,
+        ..
+    }) = footer
+    else {
         panic!("expected telemetry footer");
     };
     assert_eq!(ui.resolve(*id), "deck.track.title");
